+++ src/dialog.js
@@ -44,7 +44,10 @@ define(function(require, exports, module) {
             },
 
             // 是否有背景遮罩层
-            hasMask: true,
+            // 可设为 false
+            hasMask: {
+                hideOnClick: true
+            },
 
             // 关闭按钮可以自定义
             closeTpl: '×',
@@ -218,9 +221,13 @@ define(function(require, exports, module) {
                 if (hasMask) {
                     oldZIndex =  mask.get('zIndex');
                     mask.set('zIndex', zIndex - 1).show();
-                    mask.element.one('click', function() {
-                        that.hide();
-                    });
+                    
+                    // 点击遮罩关闭对话框
+                    if (hasMask.hideOnClick) {
+                        mask.element.one('click', function() {
+                            that.hide();
+                        });
+                    }
                 }
             });
             this.after('hide', function() {